// frontend/src/openapi/models/ConnectorTypeEnum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `langfuse` - LangFuse
 * * `langsmith` - LangSmith
 * * `braintrust` - Braintrust
 * * `galileo` - Galileo
 * * `opik` - Opik
 * * `helicone` - Helicone
 * * `arize` - Arize Phoenix
 * @export
 */
export const ConnectorTypeEnum = {
    langfuse: 'langfuse',
    langsmith: 'langsmith',
    braintrust: 'braintrust',
    galileo: 'galileo',
    opik: 'opik',
    helicone: 'helicone',
    arize: 'arize'
} as const;
export type ConnectorTypeEnum = typeof ConnectorTypeEnum[keyof typeof ConnectorTypeEnum];


export function instanceOfConnectorTypeEnum(value: any): boolean {
    for (const key in ConnectorTypeEnum) {
        if (Object.prototype.hasOwnProperty.call(ConnectorTypeEnum, key)) {
            if (ConnectorTypeEnum[key as keyof typeof ConnectorTypeEnum] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ConnectorTypeEnumFromJSON(json: any): ConnectorTypeEnum {
    return ConnectorTypeEnumFromJSONTyped(json, false);
}

export function ConnectorTypeEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): ConnectorTypeEnum {
    return json as ConnectorTypeEnum;
}

export function ConnectorTypeEnumToJSON(value?: ConnectorTypeEnum | null): any {
    return value as any;
}

export function ConnectorTypeEnumToJSONTyped(value: any, ignoreDiscriminator: boolean): ConnectorTypeEnum {
    return value as ConnectorTypeEnum;
}