// frontend/src/openapi/models/ImportPreview.ts
export interface ImportPreview {
    /**
     *
     * @type {string}
     * @memberof ImportPreview
     */
    readonly id: string;
    /**
     *
     * @type {string}
     * @memberof ImportPreview
     */
    readonly sourceProjectId: string;
    /**
     *
     * @type {Date}
     * @memberof ImportPreview
     */
    readonly windowFrom: Date | null;
    /**
     *
     * @type {Date}
     * @memberof ImportPreview
     */
    readonly windowTo: Date;
    /**
     *
     * @type {ImportPreviewStatusEnum}
     * @memberof ImportPreview
     */
    readonly status: ImportPreviewStatusEnum;
    /**
     *
     * @type {number}
     * @memberof ImportPreview
     */
    readonly traceCount: number;
    /**
     *
     * @type {number}
     * @memberof ImportPreview
     */
    readonly spanCount: number;
    /**
     *
     * @type {number}
     * @memberof ImportPreview
     */
    readonly estimatedSecondsMin: number;
    /**
     *
     * @type {number}
     * @memberof ImportPreview
     */
    readonly estimatedSecondsMax: number;
    /**
     *
     * @type {string}
     * @memberof ImportPreview
     */
    readonly error: string;
    /**
     *
     * @type {Date}
     * @memberof ImportPreview
     */
    readonly createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof ImportPreview
     */
    readonly finishedAt: Date | null;
    /**
     *
     * @type {Date}
     * @memberof ImportPreview
     */
    readonly expiresAt: Date | null;
}